Improved validation.
