+++ CCL.Creator/Validators/AudioValidator.cs
@@ -1,6 +1,7 @@
 ﻿using CCL.Types;
 using CCL.Types.Components;
 using CCL.Types.Proxies.Audio;
+using CCL.Types.Proxies.Ports;
 using UnityEngine;
 
 namespace CCL.Creator.Validators
@@ -18,6 +19,7 @@ namespace CCL.Creator.Validators
             var prefab = car.SimAudioPrefab;
 
             CheckCylCocks(prefab, result);
+            CheckPorts(prefab, result);
 
             return result;
         }
@@ -30,16 +32,33 @@ namespace CCL.Creator.Validators
                 {
                     if (audio == null)
                     {
-                        result.Fail($"{comp.name}: null entries in audio array", comp);
+                        result.Fail($"{comp.name}/{nameof(CylinderCockLayeredPortReaderProxy)}: null entries in audio array", comp);
                         return;
                     }
 
-                    if (!ComponentUtil.HasComponent<LayeredAudioProxy>(audio, false) &&
-                        !ComponentUtil.HasComponent<CopyVanillaAudioSystem>(audio, false))
+                    if (!ComponentUtil.HasComponent<LayeredAudioProxy>(audio) &&
+                        !ComponentUtil.HasComponent<CopyVanillaAudioSystem>(audio))
                     {
                         result.Warning($"{audio.name} does not have a {nameof(LayeredAudioProxy)} " +
                             $"or {nameof(CopyVanillaAudioSystem)} for {nameof(CylinderCockLayeredPortReaderProxy)}, " +
                             $"make sure this is intentional", audio);
+                    }
+                }
+            }
+        }
+
+        private void CheckPorts(GameObject prefab, ValidationResult result)
+        {
+            foreach (var comp in prefab.GetComponentsInChildren<IHasPortIdFields>(true))
+            {
+                foreach (var port in comp.ExposedPortIdFields)
+                {
+                    if (!port.IsAssigned)
+                    {
+                        if (!port.IsMultiValue && port.Required)
+                        {
+                            result.Fail($"Port field {port.FullName} must be assigned", comp is Object obj ? obj : null);
+                        }
                     }
                 }
             }
