+++ CCL.Creator/Validators/BasicValidators.cs
@@ -1,9 +1,5 @@
 ﻿using CCL.Types;
-using CCL.Types.Proxies.Controllers;
-using CCL.Types.Proxies.Controls;
 using CCL.Types.Proxies.Customization;
-using CCL.Types.Proxies.Ports;
-using CCL.Types.Proxies.Weather;
 using System.Linq;
 using UnityEngine;
 
@@ -42,7 +38,7 @@ namespace CCL.Creator.Validators
                 result.Warning($"Livery '{livery.id}' is set to spawn on a DE2 exclusive track ({exlusive.Track}), make sure this is intended", livery);
             }
 
-            if (!ComponentUtil.HasComponent<CustomizationPlacementMeshesProxy>(livery.prefab, false))
+            if (!ComponentUtil.HasComponent<CustomizationPlacementMeshesProxy>(livery.prefab))
             {
                 result.Warning($"Livery prefab does not have {nameof(CustomizationPlacementMeshesProxy)}, " +
                     "gadgets will not be able to be attached to this prefab", livery.prefab);
@@ -86,109 +82,6 @@ namespace CCL.Creator.Validators
                 return Fail($"{livery.id} - Scale is not 1");
             }
             return Pass();
-        }
-    }
-
-    [RequiresStep(typeof(LiverySettingsValidator))]
-    internal class LODGroupValidator : LiveryValidator
-    {
-        public override string TestName => "LOD Group";
-
-        protected override ValidationResult ValidateLivery(CustomCarVariant livery)
-        {
-            var result = Pass();
-            var lodGroup = livery.prefab!.GetComponent<LODGroup>();
-            if (lodGroup)
-            {
-                foreach (var lod in lodGroup.GetLODs())
-                {
-                    if (lod.renderers.Length == 0)
-                    {
-                        result.Warning($"'{livery.id}' - Missing renderers on LOD", lodGroup);
-                    }
-                }
-            }
-            return result;
-        }
-    }
-
-    [RequiresStep(typeof(LiverySettingsValidator))]
-    internal class InteriorValidator : LiveryValidator
-    {
-        public override string TestName => "Interior";
-
-        protected override ValidationResult ValidateLivery(CustomCarVariant livery)
-        {
-            var result = Pass();
-
-            if (livery.interiorPrefab != null)
-            {
-                if (livery.interiorPrefab.transform.localPosition != Vector3.zero)
-                {
-                    result.Warning($"'{livery.id}' - Interior is not centered at {Vector3.zero}", livery.interiorPrefab);
-                }
-
-                var openables = livery.interiorPrefab.GetComponentsInChildren<OpenableControlProxy>();
-
-                if (openables.Length > 0)
-                {
-                    result.Warning($"'{livery.id}' - Openables are not supported in the interior prefab", livery.interiorPrefab);
-                }
-            }
-
-            if (livery.prefab != null)
-            {
-                if (livery.prefab.transform.TryFind(CarPartNames.INTERIOR_LOD, out var interiorLOD))
-                {
-                    foreach (var item in interiorLOD.GetComponentsInChildren<Renderer>(true))
-                    {
-                        if (item.shadowCastingMode != UnityEngine.Rendering.ShadowCastingMode.Off)
-                        {
-                            result.Warning($"'{livery.id}' - Renderer {item.name} in {CarPartNames.INTERIOR_LOD} has shadow casting turned on, " +
-                                $"it should be set to off", item);
-                        }
-                    }
-                }
-                else if (livery.interiorPrefab != null)
-                {
-                    result.Warning($"{livery.id} - Interior prefab exists but {CarPartNames.INTERIOR_LOD} does not", livery.interiorPrefab);
-                }
-            }
-
-            if (livery.parentType != null && livery.parentType.HUDLayout != null && livery.interiorPrefab == null)
-            {
-                result.Fail($"{livery.id} - HUD layout needs an interior prefab to exist");
-            }
-
-            return result;
-        }
-    }
-
-    [RequiresStep(typeof(LiverySettingsValidator))]
-    internal class MultipleUnitValidator : LiveryValidator
-    {
-        public override string TestName => "Multiple Unit";
-
-        protected override ValidationResult ValidateLivery(CustomCarVariant livery)
-        {
-            if (!livery.HasMUCable) return Skip();
-
-            var result = Pass();
-            var prefab = livery.prefab!;
-
-            CheckComp<BaseControlsOverriderProxy>();
-            CheckComp<SimConnectionsDefinitionProxy>();
-            CheckComp<DamageControllerProxy>();
-
-            return result;
-            
-            void CheckComp<T>() where T : Component
-            {
-                if (!ComponentUtil.HasComponent<T>(prefab))
-                {
-                    result.Fail($"Livery '{livery.id}' has MU cable but lacks {typeof(T).Name}", livery);
-                }
-            }
         }
     }
 }
